tests/wpt/web-platform-tests/encoding-detection/en-windows-1252.tentative.html

<!doctype html>
<title>en windows-1252</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>I’ve made an encoding test.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1252", 'Expected windows-1252');
    }, "Check detection result");
      done();
};
</script>
